requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.25.1
tensorflow==2.13.0
Pillow==10.0.0
Flask-Login==0.6.2
matplotlib==3.7.2
joblib==1.3.1
